Fix: Svelte code to properly be reactive
Svelte doesn't track dependencies in function calls as seen in the
yPlusValue example
<https://svelte.dev/docs#component-format-script-3-$-marks-a-statement-as-reactive>.
In that example, total depends on both x and y. However, Svelte thinks
it only depends on x. This came to me as a surprise, because I'm used
to Vue, which does track this properly.

This means that Svelte sees classFor() depends only on col and row.
Svelte thinks classFor() does *not* depend on bfsResults or gridWorld.

To fix this, I can pass gridWorld and bfsResults explicitly as
arguments to classFor(). That way, the *call* to classFor(gridWorld,
bfsResults, …) shows the dependency explicitly, and Svelte's static
analyzer can pick up on that.

Thanks to <https://github.com/harrisi> for helping me figure this out.